Lab8_done

+++ PP_L7/src/views/user.py
@@ -4,6 +4,7 @@ from enum import Enum
 from flask import Blueprint, jsonify, request
 import src.models as models
 import src.db as db
+from src.auth import auth
 from flask_bcrypt import Bcrypt
 
 user_blueprint = Blueprint('user', __name__, url_prefix='/user')
@@ -13,105 +14,117 @@ bcrypt = Bcrypt()
 @user_blueprint.route('/', methods=['POST'])
 def create_user():
     class User(Schema):
-        email = fields.Str(required=True)
+        username = fields.Str(required=True)
         password = fields.Str(required=True)
-        firstname = fields.Str(required=True)
-        lastname = fields.Str(required=True)
-
+        firstName = fields.Str(required=True)
+        lastName = fields.Str(required=True)
+        balance = fields.Int(required=True)
 
     try:
         if not request.json:
-            raise ValidationError('Invalid input')
+            raise ValidationError('No input data provided')
         User().load(request.json)
     except ValidationError as err:
         return jsonify(err.messages), 400
 
-    new_user_model = models.Users(surname=request.json['lastname'], name=request.json['firstname'],
-                                  username=request.json['email'],
+    new_user_model = models.Users(surname=request.json['lastName'], name=request.json['firstName'],
+                                  username=request.json['username'],
                                   password=bcrypt.generate_password_hash(request.json['password']).decode('utf-8'))
     user_already_exists = db.session.query(models.Users).filter(
-        models.Users.email == new_user_model.email).count() != 0
+        models.Users.username == new_user_model.username).count() != 0
 
     if user_already_exists:
-        return jsonify({'Error': 'Duplicate user'}), 401
+        return jsonify({'ERROR': 'User already exists'}), 401
 
     try:
         db.session.add(new_user_model)
     except:
         db.session.rollback()
-        return jsonify({"Wrong user info"}), 400
+        return jsonify({"Incorrect input data"}), 400
 
-    new_user = db.session.query(models.Users).filter_by(email=request.json['email']).first()
-    new_PersonalBudget_model = models.PersonalBudgets(id=new_user_model.id, money_amount=0)
+    new_user = db.session.query(models.Users).filter_by(username=request.json['username']).first()
+    new_PersonalBudget_model = models.PersonalBudgets(id=new_user_model.id, money_amount=request.json["balance"])
     try:
         db.session.add(new_PersonalBudget_model)
     except:
         db.session.rollback()
-        return jsonify({"Database error!"}), 405
+        return jsonify({"Error with database, user not added"}), 405
 
     db.session.commit()
 
     res_json = {}
 
     res_json['id'] = new_user_model.id
-    res_json['email'] = new_user_model.email
-    res_json['firstname'] = new_user_model.firstname
-    res_json['lastname'] = new_user_model.lastname
-
-
-
+    res_json['lastName'] = new_user_model.surname
+    res_json['firstName'] = new_user_model.name
+    res_json['username'] = new_user_model.username
+    res_json['personal_budget'] = new_user_model.id
+    res_json['family_budgets'] = [int(row.family_budget_id) for row in
+                                  db.session.query(models.FamilyBudgetsUsers).filter_by(
+                                      user_id=new_user_model.id).all()]
 
     return jsonify(res_json), 200
 
 
 @user_blueprint.route('/<int:user_id>', methods=['GET'])
+@auth.login_required
 def get_user(user_id):
     user = db.session.query(models.Users).filter_by(id=user_id).first()
+    if user != auth.current_user():
+        return jsonify({'error': 'Forbidden'}), 403
     if user is None:
-        return jsonify({'Error': 'User with id not found'}), 404
+        return jsonify({'ERROR': 'No user budget with this ID'}), 404
 
     res_json = {}
 
     res_json['id'] = user.id
-    res_json['email'] = user.email
-    res_json['lastname'] = user.lastname
-    res_json['firstname'] = user.name
+    res_json['lastName'] = user.surname
+    res_json['firstName'] = user.name
+    res_json['username'] = user.username
+    res_json['pbudget'] = user.id
+    res_json['fbudgets'] = [int(row.family_budget_id) for row in
+                            db.session.query(models.FamilyBudgetsUsers).filter_by(user_id=user_id).all()]
+
     return jsonify(res_json), 200
 
 
-@user_blueprint.route('/<int:user_id>', methods=['PATCH'])
+@user_blueprint.route('/<int:user_id>', methods=['PUT'])
+@auth.login_required
 def update_user(user_id):
     try:
         class User(Schema):
-            email = fields.Str(required=True)
+            username = fields.Str(required=True)
             password = fields.Str(required=True)
-            firstname = fields.Str(required=True)
-            lastname = fields.Str(required=True)
-
-
+            firstName = fields.Str(required=True)
+            lastName = fields.Str(required=True)
 
         if not request.json:
-            raise ValidationError('Invalid id supplied')
+            raise ValidationError('No input data provided')
         User().load(request.json)
     except ValidationError as err:
         return jsonify(err.messages), 400
 
     user = db.session.query(models.Users).filter(models.Users.id == user_id).first()
+    user_n = db.session.query(models.Users).filter(models.Users.username == request.json['username']).first()
+    if user != auth.current_user():
+        return jsonify({'error': 'Unauthorized access'}), 401
     if user is None:
-        return jsonify({'Error': 'User not found'}), 404
+        return jsonify({'ERROR': 'No user with this ID'}), 404
 
     try:
-        if 'id' in request.json:
-            user.id = request.json['id']
-        if 'email' in request.json:
-            user.email = request.json['email']
-        if 'firstname' in request.json:
-            user.name = request.json['firstname']
-        if 'lastname' in request.json:
-            user.suename = request.json['lastname']
+        if user_n is not None:
+            return jsonify({'message': 'User already exists'}), 400
+        if 'username' in request.json:
+            user.username = request.json['username']
+        if 'firstName' in request.json:
+            user.name = request.json['firstName']
+        if 'password' in request.json:
+            user.password = bcrypt.generate_password_hash(request.json['password']).decode('utf-8')
+        if 'lastName' in request.json:
+            user.surname = request.json['lastName']
     except:
         db.session.rollback()
-        return jsonify({'Error': "Invalid id supplied"}), 400
+        return jsonify({'ERROR': "User data is not valid"}), 400
 
     db.session.commit()
 
@@ -119,17 +132,20 @@ def update_user(user_id):
 
 
 @user_blueprint.route('/<int:user_id>', methods=['DELETE'])
+@auth.login_required
 def delete_user(user_id):
     user = db.session.query(models.Users).filter(models.Users.id == user_id).first()
+    if user != auth.current_user():
+        return jsonify({'error': 'Unauthorized access'}), 401
     if user is None:
-        return jsonify({'Error': 'User not found'}), 404
+        return jsonify({'ERROR': 'User does not exist'}), 404
 
     try:
         db.session.delete(user)
     except:
         db.session.rollback()
-        return jsonify({'Error': "Invalid id supplied"}), 400
+        return jsonify({'ERROR': "Incorrect input data"}), 400
 
     db.session.commit()
 
-    return jsonify({'Message': "User removed"}), 200
+    return jsonify({'message': "Successfully deleted"}), 200
